requirements.txt
pyserial,flask